split query mod out of utils
